+++ system/src/Grav/Framework/Flex/FlexDirectory.php
@@ -239,9 +239,9 @@ class FlexDirectory implements FlexDirectoryInterface
 
         // If configuration is found in main configuration, use it.
         if (str_starts_with($uri, 'config://')) {
-            $path = strtr(substr($uri, 9,  -5), '/', '.');
+            $path = str_replace('/', '.', substr($uri, 9, -5));
 
-            return $grav['config']->get($path);
+            return (array)$grav['config']->get($path);
         }
 
         // Load the configuration file.
@@ -831,7 +831,7 @@ class FlexDirectory implements FlexDirectoryInterface
      * @param array $call
      * @return void
      */
-    protected function dynamicFlexField(array &$field, $property, array $call)
+    protected function dynamicFlexField(array &$field, $property, array $call): void
     {
         $params = (array)$call['params'];
         $object = $call['object'] ?? null;
@@ -840,11 +840,28 @@ class FlexDirectory implements FlexDirectoryInterface
         if ($object && method_exists($object, $method)) {
             $value = $object->{$method}(...$params);
             if (is_array($value) && isset($field[$property]) && is_array($field[$property])) {
-                $field[$property] = array_merge_recursive($field[$property], $value);
+                $value = $this->mergeArrays($field[$property], $value);
+            }
+            $field[$property] = $value;
+        }
+    }
+
+    /**
+     * @param array $array1
+     * @param array $array2
+     * @return array
+     */
+    protected function mergeArrays(array $array1, array $array2): array
+    {
+        foreach ($array2 as $key => $value) {
+            if (is_array($value) && isset($array1[$key]) && is_array($array1[$key])) {
+                $array1[$key] = $this->mergeArrays($array1[$key], $value);
             } else {
-                $field[$property] = $value;
+                $array1[$key] = $value;
             }
         }
+
+        return $array1;
     }
 
     /**