+++ CHANGELOG.md
@@ -8,7 +8,7 @@
     * Added meta support for `UploadedFile` class
     * Added support for multiple mime-types per file extension [#3422](https://github.com/getgrav/grav/issues/3422)
     * Added `setCurrent()` method to Page Collection [#3398](https://github.com/getgrav/grav/pull/3398)
-    * Initialize `$grav['uri]` before session
+    * Initialize `$grav['uri']` before session
 3. [](#bugfix)
     * Fixed `Warning: Undefined array key "SERVER_SOFTWARE" in index.php` [#3408](https://github.com/getgrav/grav/issues/3408)
     * Fixed error in `loadDirectoryConfig()` if configuration hasn't been saved [#3409](https://github.com/getgrav/grav/issues/3409)